% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suavi_brownC.R
\name{suavi_brownC}
\alias{suavi_brownC}
\title{Suavizaci칩n Brown-Cuadr치tica}
\usage{
suavi_brownC(datos, alfa)
}
\arguments{
\item{datos}{Arreglo de datos con las columnas t, Zt}

\item{alfa}{valor del parametro <1, puede ser un vector para porbar el mejor}
}
\value{
tabla de datos de los Zt suavizados y bt's suavizado y el at
}
\description{
Suavizaci칩n Brown-Cuadr치tica
}
