% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suavi_expon.R
\name{suavi_expon}
\alias{suavi_expon}
\title{Suavización Exponencial Simple}
\usage{
suavi_expon(datos, alfa)
}
\arguments{
\item{datos}{Arreglo de datos con las columnas t, Zt}

\item{alfa}{valor del parametro <1, puede ser un vector de valores}
}
\value{
tabla de datos Ft, Zt , Tt, Zt suavizado y el at
}
\description{
Método.
}
